testdisk: Autospec creation for update from version 7.1.wip to version 7.1
The package contains a GUI version of photorec: qphotorec.
To be of any practical use, this app needs to be run with root privileges,
as typically the media that needs to be recovered is removable (USB, SD card).
Therefore a separate desktop entry qphotorec-root.desktop is also created,
supplementing the original qphotorec.desktop. The new entry will
pop-up pkexec/polkit dialog asking for root password. (The original starts
with a message "this app should run with root privileges")